Add a program to launch viewers in the background.
This is an initial spike at a bundled equivalent to the mutt_bgrun
script with some extensions.

In addition to launching an external viewer program with no text output,
mutt_launch should

- close open file desriptors inherited from the main mutt process.
- clean up temp files after the viewer process closes

Only the first of these is attempted in this version and I'm not even
sure that this is the best way to do it.

I haven't changed anything in mutt_view_attachment yet.

mutt_launch: Handle closing leaked file descriptors better.

On the advice of flatcap in #93, use `getrlimit` to get the actual maximum
number of open file descriptors.  Additionally, it now checks that the file
descriptor either closes succesfully or fails with EBADF (bad file descriptor
number) or fails and exits with a message to standard error about the leaked
file descriptor.  If this causes a large number of issues we might be able to
add a few retries before exiting but I'm hesitant to do so prematurely.

mutt_launch: Send error messages to standard error.

WIP: Continuing to document attachment view-related functions.

WIP Add backgroundopen field for mailcap struct.

WIP: Opening viewers in background.